# docs/swagger.yaml
openapi: 3.0.3
info:
  title: UploadHaven API
  description: |
    UploadHaven is a modern, secure file sharing platform built with Next.js 15, TypeScript, and MongoDB. 
    This API provides comprehensive file management, real-time notifications, security scanning, and analytics capabilities.
    
    ## Features
    - 🚀 Secure file upload and sharing
    - 🔔 Real-time notifications via Server-Sent Events
    - 🔒 Malware scanning and security monitoring
    - 👥 User and admin management
    - 📊 Analytics and statistics
    - 🎯 Rate limiting and abuse prevention
    
    ## Authentication
    The API uses session-based authentication with cookies. Most endpoints require authentication.
    
  version: 2.0.0
  contact:
    name: UploadHaven Support
    url: https://github.com/yourusername/uploadhaven
    email: support@uploadhaven.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://uploadhaven.vercel.app
    description: Production server
security:
  - sessionAuth: []
...
paths:
...
  /api/auth/signup:
    post:
      tags: [Authentication]
      summary: Register a new user
      description: Create a new user account with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name:
                  type: string
                  minLength: 2
                  maxLength: 50
                  example: "John Doe"
                email:
                  type: string
                  format: email
                  example: "john@example.com"
                password:
                  type: string
                  minLength: 8
                  example: "securepass123"
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSuccess'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/signin:
    post:
      tags: [Authentication]
      summary: Sign in user
      description: Authenticate user with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: "john@example.com"
                password:
                  type: string
                  example: "securepass123"
                remember:
                  type: boolean
                  default: false
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSuccess'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/signout:
    post:
      tags: [Authentication]
      summary: Sign out user
      description: Invalidate user session and sign out
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Successfully signed out
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Successfully signed out"

  # File Management Endpoints
...
  /api/download/{shortUrl}:
    get:
      tags: [Files]
      summary: Download a file
      description: Download a file using its short URL
      parameters:
        - name: shortUrl
          in: path
          required: true
          schema:
            type: string
          description: Short URL identifier for the file
        - name: verified
          in: query
          schema:
            type: string
          description: Verification token for password-protected files
      responses:
        '200':
          description: File download
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
          headers:
            Content-Type:
              schema:
                type: string
              description: MIME type of the file
            Content-Length:
              schema:
                type: integer
              description: File size in bytes
            Content-Disposition:
              schema:
                type: string
              description: Attachment with original filename
        '404':
          description: File not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '410':
          description: File has expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Password verification required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/admin/users/{userId}/delete:
    delete:
      tags: [Admin]
      summary: Delete user
      description: Delete a user account (admin only)
      security:
        - sessionAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  deletedFiles:
                    type: integer
        '403':
          description: Admin access required
        '404':
          description: User not found
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
      description: Session-based authentication using HTTP-only cookies
  schemas:
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        details:
          type: string
...
    AuthSuccess:
      type: object
      properties:
        success:
          type: boolean
          example: true
        user:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            email:
              type: string
            role:
              type: string
              enum: [user, admin]
        message:
          type: string

    # File Types
...